requirements.txt
google-cloud-firestore==2.11.1
opencv-python-headless
functions-framework==3.3.0
google-events==0.7.0
google-api-core==2.11.1
protobuf==4.23.3
cloudevents==1.9.0